requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.6
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
cloudscraper==1.2.71
beautifulsoup4==4.12.2
pydantic==2.5.0
httpx==0.25.1